pyproject.toml
[project]
name = "advent-of-code-td"
version = "0.1.0"
description = "Advent of code solutions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-cookie3>=0.20.1",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
